Improve client readability
