+++ crates/slipstream-client/src/dns.rs
@@ -1,0 +1,396 @@
+use crate::client::ClientError;
+use crate::pacing::PacingBudgetSnapshot;
+use slipstream_core::{resolve_host_port, HostPort};
+use slipstream_dns::{build_qname, decode_response, encode_query, QueryParams, CLASS_IN, RR_TXT};
+use slipstream_ffi::picoquic::{
+    picoquic_cnx_t, picoquic_current_time, picoquic_get_path_addr, picoquic_incoming_packet_ex,
+    picoquic_prepare_packet_ex, picoquic_probe_new_path_ex, picoquic_quic_t,
+    slipstream_request_poll, PICOQUIC_PACKET_LOOP_RECV_MAX,
+};
+use slipstream_ffi::{socket_addr_to_storage, ClientConfig};
+use std::collections::HashMap;
+use std::net::{SocketAddr, SocketAddrV6};
+use tokio::net::UdpSocket as TokioUdpSocket;
+use tracing::{debug, info, warn};
+
+const PATH_PROBE_INITIAL_DELAY_US: u64 = 250_000;
+const PATH_PROBE_MAX_DELAY_US: u64 = 10_000_000;
+const DEBUG_REPORT_INTERVAL_US: u64 = 1_000_000;
+const MAX_POLL_BURST: usize = PICOQUIC_PACKET_LOOP_RECV_MAX;
+const AUTHORITATIVE_POLL_TIMEOUT_US: u64 = 5_000_000;
+
+#[derive(Clone)]
+pub(crate) struct ResolverAddr {
+    pub(crate) addr: SocketAddr,
+    pub(crate) storage: libc::sockaddr_storage,
+    pub(crate) added: bool,
+    pub(crate) probe_attempts: u32,
+    pub(crate) next_probe_at: u64,
+}
+
+pub(crate) struct DebugMetrics {
+    pub(crate) enabled: bool,
+    pub(crate) last_report_at: u64,
+    pub(crate) dns_responses: u64,
+    pub(crate) zero_send_loops: u64,
+    pub(crate) zero_send_with_streams: u64,
+    pub(crate) enqueued_bytes: u64,
+    pub(crate) send_packets: u64,
+    pub(crate) send_bytes: u64,
+    pub(crate) polls_sent: u64,
+    pub(crate) last_enqueue_at: u64,
+    pub(crate) last_report_dns: u64,
+    pub(crate) last_report_zero: u64,
+    pub(crate) last_report_zero_streams: u64,
+    pub(crate) last_report_enqueued: u64,
+    pub(crate) last_report_send_packets: u64,
+    pub(crate) last_report_send_bytes: u64,
+    pub(crate) last_report_polls: u64,
+}
+
+impl DebugMetrics {
+    pub(crate) fn new(enabled: bool) -> Self {
+        Self {
+            enabled,
+            last_report_at: 0,
+            dns_responses: 0,
+            zero_send_loops: 0,
+            zero_send_with_streams: 0,
+            enqueued_bytes: 0,
+            send_packets: 0,
+            send_bytes: 0,
+            polls_sent: 0,
+            last_enqueue_at: 0,
+            last_report_dns: 0,
+            last_report_zero: 0,
+            last_report_zero_streams: 0,
+            last_report_enqueued: 0,
+            last_report_send_packets: 0,
+            last_report_send_bytes: 0,
+            last_report_polls: 0,
+        }
+    }
+}
+
+pub(crate) struct DnsResponseContext<'a> {
+    pub(crate) quic: *mut picoquic_quic_t,
+    pub(crate) local_addr_storage: &'a libc::sockaddr_storage,
+    pub(crate) pending_polls: &'a mut usize,
+    pub(crate) inflight_poll_ids: &'a mut HashMap<u16, u64>,
+    pub(crate) debug: &'a mut DebugMetrics,
+    pub(crate) authoritative: bool,
+}
+
+pub(crate) fn resolve_resolvers(resolvers: &[HostPort]) -> Result<Vec<ResolverAddr>, ClientError> {
+    let mut resolved = Vec::with_capacity(resolvers.len());
+    for resolver in resolvers {
+        let addr = resolve_host_port(resolver).map_err(|err| ClientError::new(err.to_string()))?;
+        let addr = normalize_dual_stack_addr(addr);
+        resolved.push(ResolverAddr {
+            addr,
+            storage: socket_addr_to_storage(addr),
+            added: false,
+            probe_attempts: 0,
+            next_probe_at: 0,
+        });
+    }
+    Ok(resolved)
+}
+
+pub(crate) fn normalize_dual_stack_addr(addr: SocketAddr) -> SocketAddr {
+    match addr {
+        SocketAddr::V4(v4) => {
+            SocketAddr::V6(SocketAddrV6::new(v4.ip().to_ipv6_mapped(), v4.port(), 0, 0))
+        }
+        SocketAddr::V6(v6) => SocketAddr::V6(v6),
+    }
+}
+
+pub(crate) fn sockaddr_storage_to_socket_addr(
+    storage: &libc::sockaddr_storage,
+) -> Result<SocketAddr, ClientError> {
+    slipstream_ffi::sockaddr_storage_to_socket_addr(storage).map_err(ClientError::new)
+}
+
+pub(crate) fn expire_inflight_polls(inflight_poll_ids: &mut HashMap<u16, u64>, now: u64) {
+    if inflight_poll_ids.is_empty() {
+        return;
+    }
+    let expire_before = now.saturating_sub(AUTHORITATIVE_POLL_TIMEOUT_US);
+    let mut expired = Vec::new();
+    for (id, sent_at) in inflight_poll_ids.iter() {
+        if *sent_at <= expire_before {
+            expired.push(*id);
+        }
+    }
+    for id in expired {
+        inflight_poll_ids.remove(&id);
+    }
+}
+
+pub(crate) fn handle_dns_response(
+    buf: &[u8],
+    peer: SocketAddr,
+    ctx: &mut DnsResponseContext<'_>,
+) -> Result<(), ClientError> {
+    if let Some(response_id) = dns_response_id(buf) {
+        ctx.debug.dns_responses = ctx.debug.dns_responses.saturating_add(1);
+        if ctx.authoritative {
+            ctx.inflight_poll_ids.remove(&response_id);
+        }
+    }
+
+    if let Some(payload) = decode_response(buf) {
+        let mut peer_storage = socket_addr_to_storage(peer);
+        let mut local_storage = unsafe { std::ptr::read(ctx.local_addr_storage) };
+        let mut first_cnx: *mut picoquic_cnx_t = std::ptr::null_mut();
+        let mut first_path: libc::c_int = 0;
+        let current_time = unsafe { picoquic_current_time() };
+        let ret = unsafe {
+            picoquic_incoming_packet_ex(
+                ctx.quic,
+                payload.as_ptr() as *mut u8,
+                payload.len(),
+                &mut peer_storage as *mut _ as *mut libc::sockaddr,
+                &mut local_storage as *mut _ as *mut libc::sockaddr,
+                0,
+                0,
+                &mut first_cnx,
+                &mut first_path,
+                current_time,
+            )
+        };
+        if ret < 0 {
+            return Err(ClientError::new("Failed processing inbound QUIC packet"));
+        }
+        if !ctx.authoritative {
+            *ctx.pending_polls = ctx.pending_polls.saturating_add(1).min(MAX_POLL_BURST);
+        }
+    }
+    Ok(())
+}
+
+#[allow(clippy::too_many_arguments)]
+pub(crate) async fn send_poll_queries(
+    cnx: *mut picoquic_cnx_t,
+    udp: &TokioUdpSocket,
+    config: &ClientConfig<'_>,
+    local_addr_storage: &mut libc::sockaddr_storage,
+    dns_id: &mut u16,
+    pending_polls: &mut usize,
+    inflight_poll_ids: &mut HashMap<u16, u64>,
+    send_buf: &mut [u8],
+    debug: &mut DebugMetrics,
+) -> Result<(), ClientError> {
+    let mut remaining = *pending_polls;
+    *pending_polls = 0;
+
+    while remaining > 0 {
+        let current_time = unsafe { picoquic_current_time() };
+        unsafe {
+            slipstream_request_poll(cnx);
+        }
+
+        let mut send_length: libc::size_t = 0;
+        let mut addr_to: libc::sockaddr_storage = unsafe { std::mem::zeroed() };
+        let mut addr_from: libc::sockaddr_storage = unsafe { std::mem::zeroed() };
+        let mut if_index: libc::c_int = 0;
+        let ret = unsafe {
+            picoquic_prepare_packet_ex(
+                cnx,
+                -1,
+                current_time,
+                send_buf.as_mut_ptr(),
+                send_buf.len(),
+                &mut send_length,
+                &mut addr_to,
+                &mut addr_from,
+                &mut if_index,
+                std::ptr::null_mut(),
+            )
+        };
+        if ret < 0 {
+            return Err(ClientError::new("Failed preparing poll packet"));
+        }
+        if send_length == 0 || addr_to.ss_family == 0 {
+            *pending_polls = remaining;
+            break;
+        }
+
+        remaining -= 1;
+        *local_addr_storage = addr_from;
+        debug.send_packets = debug.send_packets.saturating_add(1);
+        debug.send_bytes = debug.send_bytes.saturating_add(send_length as u64);
+        debug.polls_sent = debug.polls_sent.saturating_add(1);
+
+        let poll_id = *dns_id;
+        let qname = build_qname(&send_buf[..send_length], config.domain)
+            .map_err(|err| ClientError::new(err.to_string()))?;
+        let params = QueryParams {
+            id: poll_id,
+            qname: &qname,
+            qtype: RR_TXT,
+            qclass: CLASS_IN,
+            rd: true,
+            cd: false,
+            qdcount: 1,
+            is_query: true,
+        };
+        *dns_id = dns_id.wrapping_add(1);
+        let packet = encode_query(&params).map_err(|err| ClientError::new(err.to_string()))?;
+
+        let dest = sockaddr_storage_to_socket_addr(&addr_to)?;
+        let dest = normalize_dual_stack_addr(dest);
+        udp.send_to(&packet, dest)
+            .await
+            .map_err(|err| ClientError::new(err.to_string()))?;
+        if config.authoritative {
+            inflight_poll_ids.insert(poll_id, current_time);
+        }
+    }
+
+    Ok(())
+}
+
+pub(crate) fn maybe_report_debug(
+    debug: &mut DebugMetrics,
+    now: u64,
+    streams_len: usize,
+    pending_polls: usize,
+    inflight_polls: usize,
+    pacing_snapshot: Option<PacingBudgetSnapshot>,
+) {
+    if !debug.enabled {
+        return;
+    }
+    if debug.last_report_at == 0 {
+        debug.last_report_at = now;
+        return;
+    }
+    let elapsed = now.saturating_sub(debug.last_report_at);
+    if elapsed < DEBUG_REPORT_INTERVAL_US {
+        return;
+    }
+    let dns_delta = debug.dns_responses.saturating_sub(debug.last_report_dns);
+    let zero_delta = debug.zero_send_loops.saturating_sub(debug.last_report_zero);
+    let zero_stream_delta = debug
+        .zero_send_with_streams
+        .saturating_sub(debug.last_report_zero_streams);
+    let enq_delta = debug
+        .enqueued_bytes
+        .saturating_sub(debug.last_report_enqueued);
+    let send_pkt_delta = debug
+        .send_packets
+        .saturating_sub(debug.last_report_send_packets);
+    let send_bytes_delta = debug
+        .send_bytes
+        .saturating_sub(debug.last_report_send_bytes);
+    let polls_delta = debug.polls_sent.saturating_sub(debug.last_report_polls);
+    let enqueue_ms = if debug.last_enqueue_at == 0 {
+        0
+    } else {
+        now.saturating_sub(debug.last_enqueue_at) / 1_000
+    };
+    let pacing_summary = if let Some(snapshot) = pacing_snapshot {
+        format!(
+            " pacing_rate={} qps_target={:.2} target_inflight={} gain={:.2}",
+            snapshot.pacing_rate, snapshot.qps, snapshot.target_inflight, snapshot.gain
+        )
+    } else {
+        String::new()
+    };
+    debug!(
+        "debug: dns+={} send_pkts+={} send_bytes+={} polls+={} zero_send+={} zero_send_streams+={} streams={} enqueued+={} last_enqueue_ms={} pending_polls={} inflight_polls={}{}",
+        dns_delta,
+        send_pkt_delta,
+        send_bytes_delta,
+        polls_delta,
+        zero_delta,
+        zero_stream_delta,
+        streams_len,
+        enq_delta,
+        enqueue_ms,
+        pending_polls,
+        inflight_polls,
+        pacing_summary
+    );
+    debug.last_report_at = now;
+    debug.last_report_dns = debug.dns_responses;
+    debug.last_report_zero = debug.zero_send_loops;
+    debug.last_report_zero_streams = debug.zero_send_with_streams;
+    debug.last_report_enqueued = debug.enqueued_bytes;
+    debug.last_report_send_packets = debug.send_packets;
+    debug.last_report_send_bytes = debug.send_bytes;
+    debug.last_report_polls = debug.polls_sent;
+}
+
+pub(crate) fn add_paths(
+    cnx: *mut picoquic_cnx_t,
+    resolvers: &mut [ResolverAddr],
+) -> Result<(), ClientError> {
+    if resolvers.len() <= 1 {
+        return Ok(());
+    }
+
+    let mut local_storage: libc::sockaddr_storage = unsafe { std::mem::zeroed() };
+    let ret = unsafe { picoquic_get_path_addr(cnx, 0, 1, &mut local_storage) };
+    if ret != 0 {
+        return Ok(());
+    }
+    let now = unsafe { picoquic_current_time() };
+
+    for resolver in resolvers.iter_mut().skip(1) {
+        if resolver.added {
+            continue;
+        }
+        if resolver.next_probe_at > now {
+            continue;
+        }
+        let mut path_id: libc::c_int = -1;
+        let ret = unsafe {
+            picoquic_probe_new_path_ex(
+                cnx,
+                &resolver.storage as *const _ as *const libc::sockaddr,
+                &local_storage as *const _ as *const libc::sockaddr,
+                0,
+                now,
+                0,
+                &mut path_id,
+            )
+        };
+        if ret == 0 && path_id >= 0 {
+            resolver.added = true;
+            info!("Added path {}", resolver.addr);
+            continue;
+        }
+        resolver.probe_attempts = resolver.probe_attempts.saturating_add(1);
+        let delay = path_probe_backoff(resolver.probe_attempts);
+        resolver.next_probe_at = now.saturating_add(delay);
+        warn!(
+            "Failed adding path {} (attempt {}), retrying in {}ms",
+            resolver.addr,
+            resolver.probe_attempts,
+            delay / 1000
+        );
+    }
+
+    Ok(())
+}
+
+fn path_probe_backoff(attempts: u32) -> u64 {
+    let shift = attempts.saturating_sub(1).min(6);
+    let delay = PATH_PROBE_INITIAL_DELAY_US.saturating_mul(1u64 << shift);
+    delay.min(PATH_PROBE_MAX_DELAY_US)
+}
+
+fn dns_response_id(packet: &[u8]) -> Option<u16> {
+    if packet.len() < 12 {
+        return None;
+    }
+    let id = u16::from_be_bytes([packet[0], packet[1]]);
+    let flags = u16::from_be_bytes([packet[2], packet[3]]);
+    if flags & 0x8000 == 0 {
+        return None;
+    }
+    Some(id)
+}
